% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service-cache.R
\name{tosi_cache_clear}
\alias{tosi_cache_clear}
\title{Clear cached table results}
\usage{
tosi_cache_clear()
}
\value{
Invisible \code{NULL}.
}
\description{
Reset both the memory and disk caches in the current R session. The next
matching \code{\link[=tosi]{tosi()}} or \code{\link[=tosi_data]{tosi_data()}} call makes a new service request.
See \code{\link[=tosi_options]{tosi_options()}} for cache size settings and lifetime limitations.
}
\examples{
tosi_cache_clear()
}
